+++ src/cfd_mesh/SimpleMeshSettings.h
@@ -142,8 +142,8 @@ public:
 
     bool m_BeamPerElementNormal;
 
-    int m_NodeOffset;
-    int m_ElementOffset;
+    unsigned long long int m_NodeOffset;
+    unsigned long long int m_ElementOffset;
 
 };
